Return NumOpResult when calculating fee
Currently we call the fee calculation functions `checked_div_by_foo` and
return an `Option` (to mimic std). However out `Amount` and `FeeRate`
are very different to std numeric types. We have the `NumOpResult`
already, lets use it.

While we are at it remove the `checked_` prefix since the return value
is now not `Option`.

Includes a bunch of changes to the `psbt` tests because extracting the
transaction from a PSBT has a bunch of overflow paths that need testing
caused by fee calculation.