units: Test generic Add and Sub
We have `Add` and `Sub` implemented for `NumOpResult<T>` but because the
impls are generic they are not grouped in the file with the other
add/sub impls. This makes it hard to see if they are supported.

Add to the `add` and `sub` unit tests to verify support is implemented.