primitives: Inline impl_to_hex_from_lower_hex
Macros are basically just a pain in the arse. This one is used to
reduced code duplication/verbosity even thought the function can be
written as a one-liner.

Also calling macros across crate boundries is error prone and a
constant source of maintenance grief.

Note also:

- The generated docs are quite general
- `#[inline]` is missing
- The docs on the macro are wrong in regards to the no-op

Remove all calls to the macro from `primitives` and inline using terse
syntax by way of the `format!` macro.