Inline to_hex instead of using macro for script types
Macros are basically just a pain in the arse. This one is used to
reduced code duplication/verbosity. There is no real other advantage.

Also calling macros across crate boundries is error prone and a
constant source of maintenance grief. Sure this one is trivial but if
we want to do script tagging it will get less trivial.

Note also:

- The docs are quite general
- `#[inline]` is missing

For the script types inline the logic and improve the docs while we
are at it.