Add script tagging
Move the script types into private modules then create a bunch of
alias', one of which is the same as the type name. This is a hack so
that we don't have to touch the million places in `bitcoin` that use
the `Script` and `ScriptBuf` - these are defined as `ContextUnknown`.

Instead, we call conversion functions at most of the call sites in
`bitcoin` with the exception of empty constructors e.g.,
`ScriptBuf::new()`. This makes it easier to review because there are
less API changes and the conversion functions can be removed later
piecemeal.