Update the doc of amounts
We've changed the API to return a result type when a numeric operation
fails but the documentation wasn't updated to reflect this. We also
added the 21M limit to the types and this was also completely
undocumented and unexplained.

This commit puts shared documentation of the numeric operations and the
limit to the `amount` module and refers to it from both types to make
sure users see it and minimize desync from copy-pasting the same
documentation to both types.