Remove amount error re-exports
The error type re-exports in the `amount` module clutter the docs. We
can just make the `error` module public and not re-export them.

This patch sets a policy of having a public `error` module and then
at the crate root re-exporting only those error types that show up
directly in the API. For the more minor error types users must got to
the `error` module.

Also, in order to control the section of the HTML docs that types show
up in, so as not to clutted the `Structs` section we manually
re-create the `amount` module in `primitives` and `bitcoin` - phew.