Remove checked_sum impl for Weight
The `Weight` type does not use `NumOpResult<Weight>` with addition,
`Add` has `Output = Weight` and uses wrapping addition. This is as
expected and as desired. We do provide `Weight::checked_add` that
returns an `Option<Weight>`, also as expected and as desired.

Further context: `Amount` was deemed 'special' such that wrapping is
considered to big a risk to have silent wrapping so we added the
`NumOpResult` type. This argument does not apply to `Weight`.

We would like to remove the `CheckedSum` trait because summing of
`Amounts` can now be done with a combination of`Sum` and the
`NumOpResult` type but `Weight` currently implements `CheckedSum`.

For the reasons outline in regards to wrapping a `Weight` type,
providing `CheckedSum` does not add much value - remove it.

Done as preparation for removing the `CheckedSum` trait all together.