Remove amount serde traits
Now that a `SignedAmount` can be infallibly converted to an `Amount`
we do not need all the internal complexity in the `amount::serde`
module, we can just de/serialize `SignedAmount` and convert it to
`Amount`. This also has the advantage if someone has a type that is
`Into<SignedAmount>` and `TryFrom<SignedAmount>` they can de/serialize
it also - BOOM!

Props to Kix for the idea.