units: Fix amount re-exports
We are introducing a new re-export policy:

For modules (e.g., `amount`):

- re-export at the module level the 'core' types for the module and
  use doc inline. 
- re-export at the module level error types found directly in the API
  and do _not_ use doc inline.
- include a public `error` module for other minor error types.

At the crate level:

- re-export only the 'core' types for the crate.

For example `Display` is 'core' to `amount` so is re-exported at
`amount::Display` but it is not 'core' to the `units` module so it is
not re-exported at the crate root.

Implement the new policy for `amount`. All other modules in `units`
are correct and need no changes re re-exports.