units: Make result module public
The `result` module provides three types, two of which are quite
low-level and do not appear directly in any APIs.

In the spirit of the new re-export policy the two minor types probably
should not be re-exported at the crate root. To implement this we need
to make the module public.