Update fuzz script to include `arbitrary` and `p2p` as a dependency
`arbitrary` is necessary to use the impls from the `bitcoin` crate's
`arbitrary` feature, and it seems like `p2p` was added to the Cargo.toml
instead of to the file generation script and then generated. The changes
to the order of the binary targets are just a result of running the
generation script again.