primitives: split ScriptSig and ScriptSigBuf from Script and ScriptBuf
Adds an extension trait ScriptSigExt with a single method, redeem_script,
which isn't used anywhere (or even tested) but seems useful enough so
we'll keep it.

Regarding extension traits: annoyingly I need to add a new extension trait
for every tagged script. In a sensible language I could just use the
existing trait and stick `where Tag == ScriptSigTag` onto the methods
that were specific to script sigs, but Rust is not sensible. It has a
fractal surface area where there is no part of the syntax that you can
ever understand all the rules of.

Anyway there is an 11-year-old tracking issue for this
https://github.com/rust-lang/rust/issues/20041