primitives: rename Script{Buf,} to GenericScript{Buf,}
...and add type aliases to the old names so that the change is invisible
outside of primitives.

Done by search-and-replace on the three files in the script/ tree, and then
reversing the search-and-replace on the tests module and the stuff related
to scripthashes. (In the next commit we will add generics to these types,
and we will want our tests and API test to continue using specific types.
Changing them to be generic would be a lot of noise, and premature. For
the script hashes, there are specific types these should be implemented
on; we don't want them to be generic even if it were possible in a nice
way.)

At the end of this PR we will undo this rename. But we want to be sure that
every single instance of `Script` and `ScriptBuf` are replaced by a more
specific type, and by using these names for `GenericScript<Whatever>`, a
"specific type" that we'll delete at the end of the PR, we can identify
all of them.