Add a `ProtocolVersion` type with named values
Protocol versions should be comparable by name when considering if we
would like to maintain a connection or not. For instance it is nice to
be able to say "reject all connections below this feature". It is also
nice to know you aren't sending some nonsense version (e.g. 70071)
because of a typo.