p2p: Add a `UserAgent` type
There is a BIP that defines what a typical user agent should look like
in a `version` message sent by a bitcoin node. All of the formatting
suggestions are just recommendations, but users that would like to
follow the recommendations should be able to do so easily. Instead of
wasting their time reading a string formatting BIP, I attempt to make it
as easy as possible here to create a user agent correctly. In practice
people put stupid shit in here all the time, so there is no validation
when reading these off the wire.

ref: https://github.com/bitcoin/bips/blob/master/bip-0014.mediawiki