p2p: Stop assuming the protocol version
The `crate::PROTOCOL_VERSION` is used implicitly in a few places.
Presumably this is for convenience, but this may lead to unexpected
results if the user advertises a particular protocol version then
implicitly uses a lower version when requesting blocks/headers etc.

I am not seeing the pain in providing the version explicitly as a user.