define_extension_trait: allow 'where T: Bound' and `where T = Type` clauses on methods
Later in this PR we are going to gate some script methods on a subtrait of their tag.
In particular, there are some methods related to sighash computation and P2SH-wrapped
segwit which require us to treat redeem scripts and witness scripts as similar things,
even though they will be different types.