primitives: split ScriptPubKey and ScriptPubKeyBuf from generic script
There are a couple places where (in Bitcoin, not just this crate) there
is confusion between scriptPubKeys and redeemScripts. Specifically, in
P2SH-wrapped segwit, we put a segwit scriptpubkey into a redeemScript
slot and then it must be treated as both.

The next commit will introduce the RedeemScript type and a trait that
covers both RedeemScript and ScriptPubKey.

Meanwhile, there are a couple methods that belong in this trait, which
in this commit I just implemented for GenericScript<T> for all T:

* in bitcoin/src/crypto/sighash.rs all the sighash methods
* in primitives/src/script/mod.rs the ScriptHash constructors
* in bitcoin/src/script/mod.rs for new_witness_program_unchecked
* in bitcoin/src/script/borrowed.rs for all the p2sh methods
* in bitcoin/src/script/owned.rs for the new_p2wpkh constructor

In the next commit we will tighten these so that they are only applicable
to the two types RedeemScript and ScriptPubKey. (Really, we want to
tighten them so that they're applicable only to RedeemScript and some
more-specific SegwitV0ScriptPubKey type. But I think that would be a
bridge too far: if we had multiple scriptpubkey types then what would
we put in TxOut::script_pubkey?)