p2p: Remove deprecated const
We backported the deprecation of this const to `v0.32.x` and even
though we typically keep deprecated things around for two releases
since `p2p` is a new crate lets just get rid of it.

FTR with this applied there are no more uses of the `deprecated`
attribute in `p2p`.