Add empty consensus_encoding crate
Add an empty `consensus_encoding` crate so we can grab the name on
crates.io

## Aims of the crate

This crate will hold the encoding traits and be depended on by
`units`, `primitives`, and `bitcoin` where implementations of the
traits will live.

All this will be feature gated so users can opt in. This nicely
bypasses the `Read`/`Write` vs `BufRead`/`BufWrite` problem because
later we can add an additional crate for the buffered stuff and users
can opt in to the one they want to use.

This crate will be no-std by default and depend on `bitcoin-io` for
I/O.

First release will just bring everything as it is from `bitcoin` over
here, then we can implement the traits in the various other crates,
then iterate - like its our job.