Implement Encodable/Decodable for list of u8
Currently we have a macro that implements for a vector of a bunch of
types, also we implement manually for `Vec<u8>`. This was done because
of the lack of specialization in Rust. However we can check the type
ID of the generic type at run time and switch on it.

Implement both traits using `TypeId` to dynamically run separate code
blocks for `u8` vs any other type.