Move CheckedData to p2p
At first blush the `CheckedData` type looks like a general purpose
checksum abstraction - it is the first four bytes of the `sha256d`
hash of a byte vector. However, in this repo it is only ever used in
`p2p` so the argument could be made that it is a checksum abstraction
defined and used by the Bitcoin p2p layer.

Furthermore the `CheckedData` type introduces a dependency on
`hashes`, we are about to separate out the consensus encoding traits
into a new crate and it would be nice to avoid a dependency on `hashes`.

In preparation for moving the consensus encoding logic into a new
crate move the `CheckedData` type to the `p2p` crate. 

If we later find the checksum abstraction useful outside of p2p we can
move or duplicate it with little maintenance burden or API bother.