Use BlockHash to test serialization behaviour
We have a test that verifies the error returned when deserializing a
vector of encodable types on the boundary when the len is equivalent
to `MAX_VEC_SIZE`. Currently the test is using a vector of
`CheckedData` structs but the exact type is immaterial to the logic of
the test. In preparation for moving the `CheckedData` type to the
`p2p` crate patch the test to use `BlockHash` instead (chosen
arbitrarily).

The exact type used to verify error returned 

I claim that this patch introduces zero test coverage change.