feat(p2p): make the p2p crate no_std
Establish the `std` feature flags on the crate and make it the default
so no change for users. This makes current and future features
explicitly std or alloc dependent.

The crate is actually pretty close to being no_std. The big tie to std
is the IP types in the std net library, but those are moved to core in
rust 1.77.

The crate is closely tied to allocation, with the upstream bitcoin crate
also requiring allocation for blocks and transactions. No alloc support
will be a larger undertaking.