Use full path for internal consensus macros
We are moving towards not calling macros across crate boundries. As
part of this we are starting to duplicate internal macros across the
stack. In an effort to make all this explicit add a single level of
path to the consensus encoding macros and import `internal_macros`.

Note that, depending on how much context you have configure to show, a
call to `crate::internal_macros::define_extension_trait` will be in
the diff of this patch. That will be changed separately next.