Invert hashes/io dependency
Currently the `io` crate depends on `hashes` and has a bunch of stuff
to add support for reading and writing to a hash engine.

We are about to create a consensus encoding crate and in order to
implement the traits for sha256 and sha256d hash types we need to
invert the dependency otherwise we get a cycle.

  io -> hashes -> consensus_encoding -> io