Invert hashes/io dependency
Currently the `io` crate depends on `hashes` and has a bunch of stuff
to add support for reading and writing to a hash engine.

While it is currently not essential _which_ way the dependency goes
there is a mild argument to be made that `io` should be a leaf crate
and `hashes` should be 'higher' up the stack.

Two considerations, both arguably quite minor:

- We can then add a (feature gated) `core2` dep to `hashes`
- It makes sense that if one wanted to hash to a `Writer` then one
  would enable an "io" feature on `hashes`. Currently instead one must
  add an `io` dependency and enable `hashes`, arguably more contrived.