Manually print InvalidChecksum error
Currently the `ParseError` is a mess. The correct fix is likely to use
an associated type in `Decodable` for the error but that can wait.

Right now we are using `hex` in the `Display` impl for `ParseError`.
This will cause trouble when we move the error type to
`consenus_encoding` because `hex` is a feature.

Just manually print hex characters by accessing the array. Logic is
already tested with a static string, this patch does not change the
output.