p2p: Allow building a "useless" address
The version message in the p2p protocol includes two `p2p::Address`
fields, one of these fields is completely ignored in practice.
Additionally, one might want to send an unreachable address if they do
not want to be discovered as a peer, i.e. they are not listening for
connections.

I opted out of using `Default` as this is a specific use case, and there
is no reasonable default for an IP address.

Reference to the line in Bitcoin Core that drops the `Address`:
https://github.com/bitcoin/bitcoin/blob/v29.1rc1/src/net_processing.cpp#L3435