Remove clippy config entry: match_on_vec_items
Clippy emits:

warning: lint `clippy::match_on_vec_items` has been removed:
`clippy::indexing_slicing` covers indexing and slicing on `Vec<_>`

IIUC one can add any just as a lint config and it just triggers a
warning and does nothing else so this usage of a removed lint config
item is essentially a no-op. Removing it makes no difference other
than removing the warning.

FTR `indexing_slicing` is not a drop in replacement because it is more
restrictive, we can use it in `units` with no other changes but not in
`primitives`.