Fix overflow during *_ceil FeeRate conversions
This fixes an overflow bug which occured when these ceil functions were
called on `FeeRate::MAX`, which is u64::MAX under the hood. This
resulted in the function panicking due to an overflow since part of the
ceil operation involves adding to the FeeRate before dividing by the
appropriate unit. By calling `saturating_add` instead of using `+`, this
prevents the function from panicking while preserving the expected
behavior when converting `FeeRate`s that would have previously
overflowed