primitives: Remove serde from hash newtypes
There are two justifications for this patch:

1. We are trimming down `serde` support in `primitives` in general,
the hash types all support consensus encoding so can be serde'd that
way.

2. The current implementation of `serde` traits for hash newtypes uses
a macro from `hashes`. In general calling macros across crate
boundaries in frowned upon in this repo. More concretely we want to
hack the hash newtypes to use inherent methods instead of implementing
`Hash` so that we can remove `hashes` from the public API of
`primitives`.

Lets just get rid of the macro calls and remove direct serde support
for hash newtypes.

JJ`: This commit contains the following changes: