p2p: Extend `BlockHash` with all zeros
The beloved `all_zeros` constructor was removed at some point. The all
zero hash does have a use case over the p2p network however. It is used
to indicate that we would like to receive the maximum amount of data
this peer has to offer. For instance, when used as the `stop_hash` when
requesting blocks, the peer will simply respond with 500 blocks.

ref: https://en.bitcoin.it/wiki/Protocol_documentation#getblocks
ref: https://github.com/bitcoin/bitcoin/blob/v29.1rc1/src/net_processing.cpp#L4047