primitives: split WitnessScript and WitnessScriptBuf from script
Ok, after the previous huge/ugly commits, this one is pretty cathartic.
I introduced the new type, used it for every field named `witness_script`
and all the constructors for `WScriptHash`, chased through all the
compiler bugs, and everything "just worked".

There is one oddity: in BIP143 there is a special-cased scriptcode construction
used for P2SH-wrapped segwit. We have methods that compute this. To make the
types line up (without introducing new traits and generics) I simply typed
the special-case thing as a WitnessScript, since that's what it's pretending
to be. I added a doccomment to all the methods that compute this.

This is a mild abuse of the `WitnessScript` type, since this special-case
object really isn't a witness script (though you could use it as one, and
trick wallets in mostly-harmless ways by doing so ... but this is a bug
in Bitcoin, not in our library!). But only a mild one, and I think it does
a *way* better job of guiding the user through the script-mangling weirdness
around sighashes and p2sh/p2wsh wrapping.

The final tag is TapScript, which will be done in the next commit. Then I
can throw away the Whatever tag, delete `Script`/`ScriptBuf`, and finally
rename `GenericScript`/`GenericScriptBuf` to those names.