primitives: delete Script and ScriptSig
This commit confirms that there is not a single untouched script instance
across the project; every one has been converted to a tagged script.

This triggers a number of doc changes which should have gone into previous
commits. I apologize for that. But it's a PITA to go back and locate exactly
which commit each doc change was supposed to go into.

In bitcoin/src/network/params.rs I changed the signet example to use a
WitnessScript. I wasn't sure if this script is interpreted as a witness
script or a scriptpubkey on Signet itself so I guessed.

In p2p/src/message.rs there is a "script" that is actually just a blob
of raw bytes (that happens to be a script) used in a "test" of the bloom
filter messages. See https://github.com/rust-bitcoin/rust-bitcoin/pull/580
where this was added for discussion about script vs bytes.

(BTW why did I let a bloom filtering PR into this project? This whole
system was a bad idea and a DoS vector. Is it even used?)