witness: rewrite the docs on several Taproot-related methods and deprecate a method
There are multiple issues with these docs:

* They repeatedly say "according to the rules of BIP XYZ". Duh. We should
  document when we're *not* following the rules.
* In the Taproot methods, we then say that we're just looking for 0x50 bytes
  but not really parsing stuff, which is untrue because we do parse the
  witness stack...
* ...and we say that the user should call Script::is_p2tr to see if the
  output is really Taproot, which is untrue according to BIP 341...
* ...and we don't even explain that the user needs to call is_p2tr on the
  *output script* and not the script being returned from this function.

Furthermore, we have two methods `tapscript` and `taproot_leaf_script`.
We say that the latter "does not return the control block". Neither one
returns the control block. We don't say how these methods are different
(or, for that matter, where to get the control block).

Also inside `tapscript` we have some assinine comment from Kix about how
the method is "broken" because of some distinction between "leaf script"
and "tapscript" that exists only in his head. We don't expose this to
our users. We shouldn't expose it to ourselves either. Clarify what (I
think) he means and promote it to a deprecation notice.