Remove coveralls
We added this a while back hoping it would improve the repo but
neither myself nor Andrew ever look at the output before ack'ing or
merging and the notification is annoying. I personally have not seen
any dev or reviewer comment on it leading to better test coverage. On
top of that we have a mutation testing bot that runs on the stable
crates.

So, while a good intention, it seems coveralls is not adding any value
to the restore at this stage - remove it.