update MSRV to 1.74 everywhere except in internals
If I update it in internals, our uses of `rust_version!` with values below
1.74 break. In particular the cond_const! macro breaks. So I will do that
in the next commit, so the diff for this one is simple.

We may want to put some work into getting a single source of truth here...
I had to change this constant in 29 files, plus internals/Cargo.toml.