Add terse opcodes for 1-16 and neg
The existing terms are unnecessarily long, we can make them more terse
with no loss of clarity.
    
Deprecate the long names and introduce short ones.

Change `OP_PUSHNUM_NEG1` to be `OP_1NEGATE` to match Bitcoin Core.

NOTE: This is a breaking change to `serde` serialization since we use
the exact letters of the consts when serializing.

Do not fix the introduced lint warnings, done next. Do, howener, fix
the roundtrip unit tests.