Add lockfile
These are not used by CI but are hopefully useful for testing with
Nix.

The CI here is old as hell. Just to get things working create a lock
file, copy it into place, and the use `cargo --locked` anyplace that
will be run by the MSRV job so it doesn't get updated.