hashes: Add bench for `sha3_256`
Following up on the addition of the SHA3-256 hash, it may be worthwhile
to strive for improvements. In my runs of the benchmark, small slices
perform poorly, indicating to me there may be something wrong with the
padding. As input size increases the hash is on the order of the slower
hashes in the library, which is acceptable for the use case. Those that
want to try to improve the hash will need this bench.