bip158: Return no match for empty query
Perhaps this is up for debate, but considering the use case of checking
for inclusions in a block, an empty query should not return `true` for any
element, as no elements exist in the query. If the user happens to have
no scripts to query for whatever reason, they will think every block is
relevant.