consensus_encoding: add decoder traits
Following the push_decode model, the Decoder trait has two methods,
`push_bytes()` and `end()`. These are two separate calls rather than
overloading a result in one to make it easer for final validation and
marshalling to happen in `end()`. `push_bytes()` operates on a mutable
reference to a slice for easier decoder composition.

Decoder is Sized since it consumes self in `end()`.