p2p: Update arbitrary dep to 1.4.1
I don't know why `cargo` can't work this out but `just
update-lock-files` is throwing an error

```
error: failed to select a version for `arbitrary`.
    ... required by package `bitcoin-p2p-messages v0.1.0 (/home/tobin/build/rust-bitcoin/p2p)`
    ... which satisfies path dependency `p2p` of package `bitcoin-fuzz v0.0.1 (/home/tobin/build/rust-bitcoin/fuzz)`
versions that meet the requirements `^1.4` are: 1.4.0

all possible versions conflict with previously selected packages.

  previously selected package `arbitrary v1.4.1`
    ... which satisfies dependency `arbitrary = "^1.4.1"` of package `bitcoin-fuzz v0.0.1 (/home/tobin/build/rust-bitcoin/fuzz)`
```

Just update the arbitrary dep to `v1.4.1` like in the other crates.