consenus_encoding: Make encoding functions use unsized
`Self` in `Encodable` is not sized (because the `Self` type within
traits is not sized by default). And also generic types as function
arguments are sized by default so the args to our encoding functions
`(e.g., `encode_to_vec`) are sized but they do not need to be.

Add explicit `?Sized` to the encoding functions.