consensus_encoding: collapse decoder errors
Adds a DecoderError trait with a method to collapse the error. For
"leaf" types like UnexpectedEof, this is simply an identity function
returning self. But for the Either type, it allows a recursive collapse
of an arbitrarily nested type, assuming the leafs all have the same
type like Either<Either<UnexpectedEof, UnexpectedEof>, UnexpectedEofi> ->
UnexpectedEof.