Remove current index from SliceEncoder
At one stage during development, IIRC, I was trying to mutate the
encoder from within a loop and was unable to do so so I added a
current index instead. With Andrew's later improved version of
`advance` this was no longer the case but I left the `cur_idx`
variable in there. Remove it.

This was noticed during review but kindly left for a follow up since
its an internal change.