primitives: Add hash type name to Debug output
The number of hash wrapper types has reached a point where
distinguishing the exact type for debugging would better serve
users.

Adjusts the generic hashtype logic to provide a macro for Debug
implementation when not using the hex feature, and a macro for all
expected fmt traits Debug, Display, LowerHex, and UpperHex) when
using the hex feature.