Check position value when decoding cursor
If a `Witness` is created on a 16-bit machine then the elements vector
should be less the `u16::MAX` and all the encode positions should fit
in a `u16` even though we leave 4 bytes for them. Since we are in a
function that returns an `Option` anyways lets add a check just to be
defensive instead of casting (wrap) and causing debug hell for some
poor user.