primitives: Add private cast_to_usize_if_valid function
I have a bit of confusion around this 4,000,000 value so add a
_private_ function to the `witness` module and call it in two places
to remove casts of `u32` to `usize` that would likely never be hit
anyways.

I also remove the issue link, the issue is till open but its not
useful now to have a link to it in code.