backport part of 4915: Deprecate weight constructors
Deprecation in trait methods is painful and these two weight
constructors are now defined in `TxInExt`.

In #4915 we deprecate `Weight::from_non_witness_data_size` and
`Weight::from_witness_data_size` in favour of `Weight::from_vb` and
`Weight::from_wu` respectively but doing so messes with the APIs of
these to `TxIn` methods because the return value changes from `Weight`
to `Option<Weight>`.

Deprecate the functions here so we can either break the API in 0.33 or
delete the functions entirely.

Note this only does the stuff from 4915 in `bitcoin` not in `units`.