Remove BytesEncoder::without_length_prefix
This constructor is a bit confusing because it means when seeing the
`BytesEncoder` in nested encoders it is not obvious if there is a
length prefix or not.

We can remove this ambiguity by only ever using an `ArrayEncoder` when
there is not a length prefix. The cost is a tiny wee hit to
performance because we may have to copy bytes around due to the
`Encodable` trait definition. As we do for `Txid` inside an
`OutPointEncoder`.

This complexity was observed by jrakibi and described in #5097. This
is an alternate to his proposed solution.