Add in diff cargo-mutants CI workflow
This adds a new CI workflow in the rust.yml workflow file to be run
alongside tests on PRs and pushes.

This CI utilizes the `--in-diff` flag on cargo-mutants to check for
mutations in the source code within the PR diff.

This does not replace the current weekly cargo-mutants CI as changes
outside of the source code, whether through a change of a test or some
other change not explicitly in the rust source the diff mutation will
not catch.