consensus_encoding: rename min_bytes_needed
The `min_bytes_needed` name is a bit of a misnomer because it implies a
lower bound instead of an upper. I believe the original intention was
to capture "minimum bytes needed to advance the state of the decoder",
but that is a little wordy. Updated to `read_limit` since this is an
upper bound to protect against over consuming bytes the decoder doesn't
know what to do with.