bitcoin: reject ambiguous merkle trees in MerkleNode::calculate_root
CVE 2012-2459 is about an ambiguity in Bitcoin's merkle root calculation
used to embed transactions in blocks. It is possible to take a valid
block, duplicate some sets of transactions, and obtain an invalid block
with the same hash. This can confuse nodes into incorrectly rejecting
the valid block.

Instead, following Core, when we detect duplicated transactions that
would cause this bug, we just refuse to return the Merkle root. Then
there is nothing that the node can get confused by; it will reject the
bad block (which is definitely bad, since it has duplicate transactions)
and move on with its life.

This patch is really just three lines (well, one line in an if guard)
plus a bunch of comments.

Fixes #5023

Thanks to Github user tm0, who responsibly disclosed this issue (though
I decided it was outside of our security boundary and requested that
they open a public issue).