primitives: remove a bunch of panics from Transaction::decoder
The state machine logic in TransactionDecoder is a bit frustrating. We
need to obtain a mutable reference to the current sub-decoder to attempt
to feed data to it. If this fails, we then need to move the sub-decoder,
attempt to end() it, then replace it if this succeeds. (And the replacement
uses the return value from end() so we can't reorder this.)

Having the end() between the move and replacement means we need a dummy
Transitioning state, and having a mutable borrow before the move means
we need some borrowck trickery.

This commit separates the mutable-borrow logic from the move logic to
avoid the trickery. It retains the Transitioning dummy state which I can't
find a way to avoid.

Refactor only; no observable behavior changes.